Move MockStream and MockSocket into their own files
